robo-aspirador/requirements.txt
pybullet>=3.2.0
numpy>=1.26.0
pyyaml>=6.0
requests>=2.31.0
matplotlib>=3.8.0
scipy>=1.11.0
opencv-python>=4.8.0
